modularize infra code a bit
